+++ template.py
@@ -22,7 +22,7 @@ list_of_files=[
     f"src/{project_name}/pipelines/training_pipeline.py",
     f"src/{project_name}/pipelines/prediction_pipeline.py",
     f"src/{project_name}/exceptions.py",
-    f"src/{project_name}/logger.py",
+    f"src/{project_name}/loggers.py",
     f"src/{project_name}/utils.py",
 
     # main files
